third_party/blink/web_tests/platform/generic/fast/frames/adopt-object-into-itself-expected.txt

ALERT: HierarchyRequestError: Failed to execute 'adoptNode' on 'Document': The node provided is a frame which contains this document.
